data-analyst-agent-openai/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai==1.3.7
requests==2.31.0
pandas==2.1.4
beautifulsoup4==4.12.2
lxml==4.9.3
matplotlib==3.8.2
scikit-learn==1.3.2
numpy==1.26.2
gunicorn==21.2.0